Makefile: move sysadmin side effects out of `make install`
The install recipe is overloaded with unrelated actions which make no
sense in an isolated, reproducible build environment.

For example, it tries to reload a systemd service manager
if the systemctl binary was present, or to set file capabilities.
Neither are possible in an isolated package build environment: there is
no service manager to reload, and there are insufficient privileges to
setcap +ep anything, so this is not only nonsensical, but fails `make
install`. Both of these actions pertain to system administration, not
release build or packaging; they are useful on a real running system or
in a development workflow.

We introduce a new target, `make install-live`, to do what `make install`
previously did on a live system, and move the setcap and systemctl
invocations to this target.

The result of this, to put it briefly, is that the following command
just works without privilege escalation:
    DESTDIR="$buildroot" make install
as it should.

Fixes: d953a6106d12 ("Improve Systemd Units for all distros (#95)")